Move visitor to data structures
